% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_abcd_flags_bonds.R
\name{fake_abcd_flags_bonds}
\alias{fake_abcd_flags_bonds}
\title{Create an example \code{abcd_flags_bonds} object}
\usage{
fake_abcd_flags_bonds(
  credit_parent_id = "000BGD-E",
  has_asset_level_data = TRUE,
  has_ald_in_fin_sector = TRUE,
  sectors_with_assets = "Power + Oil&Gas"
)
}
\arguments{
\item{credit_parent_id}{value/s to be used for the \code{credit_parent_id} column}

\item{has_asset_level_data}{value/s to be used for the \code{has_asset_level_data}
column}

\item{has_ald_in_fin_sector}{value/s to be used for the
\code{has_ald_in_fin_sector} column}

\item{sectors_with_assets}{value/s to be used for the \code{sectors_with_assets}
column}
}
\value{
A data frame with the specified columns and/or their default values
}
\description{
This function creates an example \code{abcd_flags_bonds} object.
}
