% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_abcd_scenario_bonds.R
\name{fake_abcd_scenario_bonds}
\alias{fake_abcd_scenario_bonds}
\title{Create an example \code{abcd_scenario_bonds} object}
\usage{
fake_abcd_scenario_bonds(
  scenario_source = "ISF2021",
  scenario = "NZE",
  id_name = "ar_company_id",
  id = "10015",
  ald_company_sector_id = "4",
  equity_market = "DevelopedMarket",
  scenario_geography = "Global",
  ald_sector = "Coal",
  technology = "Coal",
  year = 2021,
  plan_tech_prod = 0,
  plan_br_wt_factor = NA_real_,
  plan_br_wt_techshare = NA_real_,
  plan_emission_factor = NA_real_,
  scen_tech_prod = 0,
  scen_br_wt_factor = NA_real_,
  scen_br_wt_techshare = NA_real_,
  scen_emission_factor = 0,
  current_plan_row = 1,
  scenario_exists = 1
)
}
\arguments{
\item{scenario_source}{value/s to be used for the \code{scenario_source} column}

\item{scenario}{value/s to be used for the \code{scenario} column}

\item{id_name}{value/s to be used for the \code{id_name} column}

\item{id}{value/s to be used for the \code{id} column}

\item{ald_company_sector_id}{value/s to be used for the
\code{ald_company_sector_id} column}

\item{equity_market}{value/s to be used for the \code{equity_market} column}

\item{scenario_geography}{value/s to be used for the \code{scenario_geography}
column}

\item{ald_sector}{value/s to be used for the \code{ald_sector} column}

\item{technology}{value/s to be used for the \code{technology} column}

\item{year}{value/s to be used for the \code{year} column}

\item{plan_tech_prod}{value/s to be used for the \code{plan_tech_prod} column}

\item{plan_br_wt_factor}{value/s to be used for the \code{plan_br_wt_factor}
column}

\item{plan_br_wt_techshare}{value/s to be used for the \code{plan_br_wt_techshare}
column}

\item{plan_emission_factor}{value/s to be used for the \code{plan_emission_factor}
column}

\item{scen_tech_prod}{value/s to be used for the \code{scen_tech_prod} column}

\item{scen_br_wt_factor}{value/s to be used for the \code{scen_br_wt_factor}
column}

\item{scen_br_wt_techshare}{value/s to be used for the \code{scen_br_wt_techshare}
column}

\item{scen_emission_factor}{value/s to be used for the \code{scen_emission_factor}
column}

\item{current_plan_row}{value/s to be used for the \code{current_plan_row} column}

\item{scenario_exists}{value/s to be used for the \code{scenario_exists} column}
}
\value{
A data frame with the specified columns and/or their default values
}
\description{
This function creates an example \code{abcd_scenario_bonds} object.
}
