% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_currencies.R
\name{fake_currencies}
\alias{fake_currencies}
\title{Create an example \code{currencies} object}
\usage{
fake_currencies(currency = "USD", exchange_rate = 1)
}
\arguments{
\item{currency}{value/s to be used for the \code{currency} column}

\item{exchange_rate}{value/s to be used for the \code{exchange_rate} column}
}
\value{
A data frame with the specified columns and/or their default values
}
\description{
This function creates an example \code{currencies} object.
}
