% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_intermediate_scenario_data.R
\name{fake_intermediate_scenario_data}
\alias{fake_intermediate_scenario_data}
\title{Create an example \code{intermediate_scenario_data} object}
\usage{
fake_intermediate_scenario_data(
  source = "WEO2022",
  scenario = "NZE_2050",
  scenario_geography = "Global",
  sector = "Power",
  technology = "RenewablesCap",
  indicator = "Capacity: installed",
  units = "GW",
  year = 2025,
  value = 500
)
}
\arguments{
\item{source}{value/s to be used for the \code{source} column}

\item{scenario}{value/s to be used for the \code{scenario} column}

\item{scenario_geography}{value/s to be used for the \code{scenario_geography} column}

\item{sector}{value/s to be used for the \code{sector} column}

\item{technology}{value/s to be used for the \code{technology} column}

\item{indicator}{value/s to be used for the \code{indicator} column}

\item{units}{value/s to be used for the \code{units} column}

\item{year}{value/s to be used for the \code{year} column}

\item{value}{value/s to be used for the \code{value} column}
}
\value{
A data frame with the specified columns and/or their default values
}
\description{
This function creates an example \code{intermediate_scenario_data} object.
}
