% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacta.data.validation-package.R
\docType{package}
\name{pacta.data.validation-package}
\alias{pacta.data.validation}
\alias{pacta.data.validation-package}
\title{pacta.data.validation: pacta.data.validation}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

For more information visit \url{https://rmi.org/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RMI-PACTA/pacta.data.validation}
  \item Report bugs at \url{https://github.com/RMI-PACTA/pacta.data.validation/issues}
}

}
\author{
\strong{Maintainer}: CJ Yetman \email{cj@cjyetman.com} (\href{https://orcid.org/0000-0001-5099-9500}{ORCID}) [contractor]

Authors:
\itemize{
  \item Jackson Hoffart \email{jackson.hoffart@gmail.com} (\href{https://orcid.org/0000-0002-8600-5042}{ORCID}) [contractor]
  \item Jacob Kastl \email{jacob.kastl@gmail.com} [contractor]
  \item Alex Axthelm \email{aaxthelm@rmi.org} (\href{https://orcid.org/0000-0001-8579-8565}{ORCID}) [contractor]
}

Other contributors:
\itemize{
  \item Rocky Mountain Institute \email{PACTA4investors@rmi.org} [copyright holder, funder]
}

}
\keyword{internal}
