% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_abcd_scenario_equity.R
\name{validate_abcd_scenario_equity}
\alias{validate_abcd_scenario_equity}
\title{Validate a \code{abcd_scenario_equity} object}
\usage{
validate_abcd_scenario_equity(data)
}
\arguments{
\item{data}{An object (typically a data frame)}
}
\value{
\code{TRUE} if the object is valid, otherwise an error with a message
explaining the failed assertions
}
\description{
This function validates that an object is a valid \code{abcd_scenario_equity} dataset.
}
