% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_currencies.R
\name{validate_currencies}
\alias{validate_currencies}
\title{Validate a \code{currencies} object}
\usage{
validate_currencies(data)
}
\arguments{
\item{data}{An object (typically a data frame)}
}
\value{
\code{TRUE} if the object is valid, otherwise an error with a message
explaining the failed assertions
}
\description{
This function validates that an object is a valid \code{currencies} dataset.
}
