% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_intermediate_scenario_output.R
\name{validate_intermediate_scenario_output}
\alias{validate_intermediate_scenario_output}
\title{Validate an intermediate scenario object}
\usage{
validate_intermediate_scenario_output(data)
}
\arguments{
\item{data}{An object (typically a data frame)}
}
\value{
\code{TRUE} if the object is valid, otherwise an error with a message
explaining the failed assertions
}
\description{
This function validates that an object is a valid intermediate scenario
dataset (e.g. \code{weo_2022}, \code{geco_2022}, or more generally, \code{publication_YYYY}).
}
