% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_masterdata_debt_datastore.R
\name{validate_masterdata_debt_datastore}
\alias{validate_masterdata_debt_datastore}
\title{Validate a \code{masterdata_debt_datastore} object}
\usage{
validate_masterdata_debt_datastore(data)
}
\arguments{
\item{data}{An object (typically a data frame)}
}
\value{
\code{TRUE} if the object is valid, otherwise an error with a message
explaining the failed assertions
}
\description{
This function validates that an object is a valid \code{masterdata_debt_datastore}
dataset.
}
