% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_masterdata_ownership_ai.R
\name{validate_masterdata_ownership_ai}
\alias{validate_masterdata_ownership_ai}
\title{Validate a \code{masterdata_ownership_ai} object}
\usage{
validate_masterdata_ownership_ai(data)
}
\arguments{
\item{data}{An object (typically a data frame)}
}
\value{
\code{TRUE} if the object is valid, otherwise an error with a message
explaining the failed assertions
}
\description{
This function validates that an object is a valid \code{masterdata_ownership_ai}
dataset.
}
